% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cargar_rscript.R
\name{cargar_rscript}
\alias{cargar_rscript}
\title{Cargar R Scritp en blanco para iniciar proyecto}
\usage{
cargar_rscript(modulo = "proj_rscript")
}
\arguments{
\item{modulo}{nombre del archivo .R que será cargado}
}
\description{
Esta función crea un R Script conteniendo la estructura mínima requerida para mantener ordenada la información de un proyecto o análisis.
}
\examples{
\dontrun{
  cargar_rscript("proj_rscript")
}

}
