% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_permu_rda.R
\name{check_permu_rda}
\alias{check_permu_rda}
\title{Revisar Estadísticas de Permutaciones (ANOVA) para RDA/CCA}
\usage{
check_permu_rda(RDA, by = NULL, plot = TRUE, plot_args = list())
}
\arguments{
\item{RDA}{Un objeto resultante de `rda`, `cca` o `capscale` (vegan).}

\item{by}{(Opcional) Argumento a pasar a `anova.cca` (por defecto NULL). Puede ser "terms", "axis", "margin", etc.}

\item{plot}{(Opcional) Lógico. Si TRUE (por defecto), se muestra el gráfico de densidad. Si FALSE, no se genera.}

\item{plot_args}{(Opcional) Lista de argumentos adicionales para `densityplot`. Por defecto `list()`.}
}
\value{
Invisiblemente, una lista con los siguientes elementos:
\itemize{
  \item{Anova}{Resultado del ANOVA (anova.cca).}
  \item{Permustats}{Objeto resultante de `permustats`.}
  \item{SummaryPermustats}{Resumen del objeto Permustats.}
}
}
\description{
Esta función realiza un ANOVA para un objeto RDA o CCA (o similar, como capscale) usando `anova.cca`.
Además, muestra las estadísticas de permutaciones a través de `permustats`, su resumen, y opcionalmente
genera un gráfico de densidad de las distribuciones de las permutaciones.
}
\examples{
\dontrun{
# Suponiendo que tienes un objeto RDA (por ejemplo my_rda)
# check_permu(my_rda, by = "terms", plot = TRUE)
}

}
