% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterx_palettes.R
\name{masterx_pal}
\alias{masterx_pal}
\title{Paletas de colores masterX}
\usage{
masterx_pal(palette = c("pairs14", "pal22", "pal22_ordered"))
}
\arguments{
\item{palette}{Nombre de la paleta de colores}
}
\value{
la paleta de colores creada como una función útil en códigos de ggplot2
}
\description{
Esta función contiene las paletas de color de masterX
    para ser utilizadas en visualización de datos con ggplot2. El nombre
    de cada paleta contiene la cantidad de colores discretos almacenados
    en ella.
}
