% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterx_rstudio_theme.R
\name{masterx_rstudio_theme}
\alias{masterx_rstudio_theme}
\title{Instalación del tema masterX Rstudio}
\usage{
masterx_rstudio_theme(URL_tema)
}
\arguments{
\item{URL_tema}{Enlace URL raw del tema a aplicar. Como ejemplo, utiliza:
Enlace del tema masterX-RStudio.rstheme:
"https://raw.githubusercontent.com/irwingss/masterX-RStudio-Theme/main/masterX-RStudio.rstheme"}
}
\description{
Utiliza esta función para instalar el tema de colores de RStudio: masterX. No es requerido el uso de argumentos en la función. Al ejecutarla como `masterx_rstudio_theme()` ya procederá a descargar, instalar y aplicar el tema de manera automática. Si quieres utilizar esta función para instalar otros temas desde cualquier repositorio de GitHub, asegúrate que el tema tenga extensión .rstheme
}
\examples{
masterx_rstudio_theme(
  "https://raw.githubusercontent.com/irwingss/masterX-RStudio-Theme/main/masterX-RStudio.rstheme"
)

}
