% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_masterx.R
\name{scale_color_masterx}
\alias{scale_color_masterx}
\title{Coloreo con las paletas de colores masterX}
\usage{
scale_color_masterx(..., palette = "pal22")
}
\arguments{
\item{palette}{Nombre de la paleta de colores.}
}
\value{
la paleta de colores que será aplicada al gráfico
}
\description{
Función afín a ggplot2 para el coloreo de geométricas.
     Para conocer las paletas disponibles revisa la ayuda de la función
     masterx_palettes().
}
\examples{
library(ggplot2)
data(iris)
ggplot(data = iris,
aes(x = Petal.Length, y = Sepal.Length, color = Species)) +
  geom_point(size = 4, alpha = 0.4) +
  theme_bw() +
  theme(legend.position = "top") +
  labs(x = "Petal Length", y = "Sepal Length") +
  scale_color_masterx(palette = "pal22")
}
