% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_GBIF_thining_kuenm.R
\name{sp_GBIF_thining_kuenm}
\alias{sp_GBIF_thining_kuenm}
\title{Descargar y visualizar puntos georreferenciados de presencia de especies desde GBIF}
\usage{
sp_GBIF_thining_kuenm(
  Especie,
  ocu_reg_limite = 5000,
  separacion_puntos = 10,
  semilla = 123,
  porc_entrenamiento = 0.75,
  limpieza = TRUE,
  mapview = TRUE,
  subcarpeta_sp = NULL,
  CRS = sp::CRS("+proj=longlat +ellps=WGS84 +datum=WGS84")
)
}
\arguments{
\item{Especie}{string con el nombre científico de la especie de interés.}

\item{ocu_reg_limite}{valor numérico que especifíca la cantidad máxima (límite) de registros de ocurrencia descargados desde GBIF para la especie de interés.}

\item{separacion_puntos}{valor numérico para definir la cantidad de kilómetros a los que deben estar separados los registros de presencia como mínimo (aplicado para el procedimiento de thining).}

\item{semilla}{valor numérico, define una semilla para la selección.}

\item{porc_entrenamiento}{valor numérico, que expresa la proporción (0 a 1) de registros destinados a ser guardados en la base de entrenamiento.}

\item{limpieza}{lógico, TRUE por defecto. Activa el proceso de thining de la
base de datos y exporta tres bases de datos (base completa, la de
entrenamiento con 75\% de los registros, y la de testeo con el 25\% restante).}

\item{mapview}{lógico, TRUE por defecto. Activa la visualización de los registros completos descargados y limpiados  con mapview en RStudio.}

\item{subcarpeta_sp}{NULL por defecto. String que determina la subcarpeta donde se guardarán los archivos .csv de ocurrencia para la especie de interés.}

\item{CRS}{Sistemas de Referencia de Coordenadas. Definido por defecto como "+proj=longlat +ellps=WGS84 +datum=WGS84".}
}
\value{
La función devuelve los archivos de registros con los nombres necesarios para ser usados con el paquete Kuenm (Cobos et al., 2019) con los nombres "Sp_joint.csv", "Sp_train.csv" y "Sp_test.csv".
}
\description{
Accede al repositorio de GBIF (Global Biodiversity Information
     Facility) y realiza la descarga de los registros georreferenciados de
     presencia de la especie solicitada. La base de datos descargada puede
     ser utilizada para el modelamiento de distribución potencial de la especie.
     Además de descargar los datos, la función realiza las siguientes acciones:
\itemize{
 \item Procedimiento 2: Eliminar registros duplicados.
 \item Procedimiento 3: Eliminar registros con coordenadas (0, 0).
 \item Procedimiento 4: Guardar los archivos .csv de los registros de ocurrencia en la carpeta de trabajo actual con el nombre "*Especie*_occ_full.csv".
 \item Procedimiento 5: Reducción de puntos cercanos (Thining), con la separación mínima entre puntos definida por el argumento **separacion_puntos**. Este proceso está activado por defecto.
 \item Procedimiento 6: Guardar los archivos .csv de los registros de ocurrencia completos, y las bases de datos de entrenamiento y testeo en la carpeta de trabajo actual, con los nombres "Sp_joint.csv", "Sp_train.csv" y "Sp_test.csv", respectivamente.
 \item Creación de las bases de entrenamiento (75\% de los registros) y testeo (25\% de los registros). Este proceso está activado por defecto.
}
}
\details{
Confirmación de los nombres científicos:
Para corroborar que el nombre científico es correcto, y previsualizar la cantidad de registros disponibles para la especie, visita y realiza la búsqueda en: https://www.gbif.org/
}
\examples{
# Descarga con la configuración por defecto
sp_GBIF_thining_kuenm("Rupicola peruvianus")

# Descargar los archivos .csv de registros hacia la subcarpeta de la especie
sp_GBIF_thining_kuenm(Especie = "Syndactyla ruficollis",
subcarpeta_sp = "Syndactyla ruficollis/")

}
