% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp_thining_kuenm.R
\name{sp_thining_kuenm}
\alias{sp_thining_kuenm}
\title{Reducción de registros cercanos (thinning) para SDM}
\usage{
sp_thining_kuenm(
  data_ocurrencias,
  lat_col = "decimalLatitude",
  long_col = "decimalLongitude",
  especie_col = "scientificName",
  subcarpeta_sp = NULL,
  distancia_km = 10,
  semilla = 123,
  porc_entrenamiento = 0.75,
  mapview = TRUE,
  CRS = sp::CRS("+proj=longlat +ellps=WGS84 +datum=WGS84")
)
}
\arguments{
\item{data_ocurrencias}{data.frame con los registros de ocurrencia georreferenciados (con las columna de latitud y longitud por separado).}

\item{subcarpeta_sp}{NULL por defecto. String que determina la subcarpeta donde se guardarán los archivos .csv de ocurrencia para la especie de interés.}

\item{distancia_km}{valor numérico para definir la cantidad de kilómetros a los que deben estar separados los registros de presencia como mínimo (aplicado para el procedimiento de thining).}

\item{semilla}{valor numérico, define una semilla para la selección.}

\item{porc_entrenamiento}{valor numérico, que expresa la proporción (0 a 1) de registros destinados a ser guardados en la base de entrenamiento.}

\item{mapview}{lógico, TRUE por defecto. Activa la visualización de los registros completos descargados y limpiados  con mapview en RStudio.}

\item{CRS}{Sistemas de Referencia de Coordenadas. Definido por defecto como "+proj=longlat +ellps=WGS84 +datum=WGS84".}
}
\value{
La función devuelve los archivos de registros con los nombres necesarios para ser usados con el paquete Kuenm (Cobos et al., 2019) con los nombres "Sp_joint.csv", "Sp_train.csv" y "Sp_test.csv".
}
\description{
Recibe una data.frame de puntos georreferenciados de ocurrencia, preferiblemente descargados con la función rgbif::occ_data(), que haya pasado por un proceso preliminar de limpieza de registros con baja confiabilidad (basado en opinión de experto). En el marco del pre-procesamiento de datos para modelamiento de distribución potencial de especies propuesto por Cobos et. al (2019) en el paquete `kuenm`, esta función realiza los siguientes procedimientos:
\itemize{
 \item Procedimiento 1: Reducción de puntos cercanos (Thinning), con la separación mínima entre puntos definida por el argumento distancia_km. Este proceso está activado por defecto.
 \item Procedimiento 2: Guardar los archivos .csv de los registros de ocurrencia completos, y las bases de datos de entrenamiento y testeo en la carpeta de trabajo actual, con los nombres "Sp_joint.csv", "Sp_train.csv" y "Sp_test.csv", respectivamente.
 \item Creación de las bases de entrenamiento (75\% de los registros) y testeo (25\% de los registros). Este proceso está activado por defecto.
}
}
