% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpiar_fil_col_wcero.R
\name{limpiar_fil_col_wcero}
\alias{limpiar_fil_col_wcero}
\title{Limpieza de Filas o Columnas con Sumatoria Cero}
\usage{
limpiar_fil_col_wcero(data, by_row = TRUE, tolerance = 1e-08)
}
\arguments{
\item{data}{Una matriz o data frame a limpiar.}

\item{by_row}{Lógico. Si es \code{TRUE}, se eliminarán filas; si es \code{FALSE}, se eliminarán columnas. Por defecto es \code{TRUE}.}

\item{tolerance}{Umbral de tolerancia para considerar una sumatoria como cero. Por defecto es \code{1e-8}.}
}
\value{
Un objeto de la misma clase que \code{data} con las filas o columnas especificadas eliminadas.
}
\description{
Esta función elimina filas o columnas de una matriz o data frame cuya sumatoria es igual a cero
(dentro de una tolerancia especificada). Es útil para limpiar datos antes de realizar análisis multivariados.
}
\examples{
# Crear una data frame de ejemplo
df <- data.frame(
  A = c(1, 0, 3),
  B = c(0, 0, 0),
  C = c(4, 5, 6)
)

# Eliminar columnas con sumatoria cero
limpiar_fc_cero(df, by_row = FALSE)

# Eliminar filas con sumatoria cero
limpiar_fc_cero(df, by_row = TRUE)

}
